Refactor GDS var length function
