+++ src/function/gds/output_writer.cpp
@@ -59,7 +59,7 @@ PathsOutputWriter::PathsOutputWriter(main::ClientContext* context, RJOutputs* rj
     if (info.writeEdgeDirection) {
         directionVector = createVector(LogicalType::LIST(LogicalType::BOOL()), mm);
     }
-    lengthVector = createVector(LogicalType::INT64(), mm);
+    lengthVector = createVector(LogicalType::UINT16(), mm);
     if (info.writePath) {
         pathNodeIDsVector = createVector(LogicalType::LIST(LogicalType::INTERNAL_ID()), mm);
         pathEdgeIDsVector = createVector(LogicalType::LIST(LogicalType::INTERNAL_ID()), mm);
@@ -73,6 +73,11 @@ static void addListEntry(ValueVector* vector, uint64_t length) {
     vector->setValue(0, entry);
 }
 
+static void setLength(ValueVector* vector, uint16_t length) {
+    KU_ASSERT(vector->dataType.getLogicalTypeID() == LogicalTypeID::UINT16);
+    vector->setValue<uint16_t>(0, length);
+}
+
 void PathsOutputWriter::write(processor::FactorizedTable& fTable, nodeID_t dstNodeID) const {
     auto output = rjOutputs->ptrCast<PathsOutputs>();
     auto& bfsGraph = output->bfsGraph;
@@ -84,7 +89,7 @@ void PathsOutputWriter::write(processor::FactorizedTable& fTable, nodeID_t dstNo
         if (sourceNodeID == dstNodeID) {
             // We still output a path from src to src if required path length is 0.
             // This case should only run for variable length joins.
-            lengthVector->setValue<int64_t>(0, 0);
+            setLength(lengthVector.get(), 0);
             if (info.writePath) {
                 beginWritePath(0);
             }
@@ -92,7 +97,7 @@ void PathsOutputWriter::write(processor::FactorizedTable& fTable, nodeID_t dstNo
         }
         return;
     }
-    lengthVector->setValue<int64_t>(0, firstParent->getIter());
+    setLength(lengthVector.get(), firstParent->getIter());
     std::vector<ParentList*> curPath;
     curPath.push_back(firstParent);
     auto backtracking = false;
@@ -127,7 +132,7 @@ void PathsOutputWriter::write(processor::FactorizedTable& fTable, nodeID_t dstNo
                 curPath[curPath.size() - 1] = top->getNextPtr();
                 backtracking = false;
                 if (curPath.size() == 1) {
-                    lengthVector->setValue<int64_t>(0, curPath[0]->getIter());
+                    setLength(lengthVector.get(), curPath[0]->getIter());
                 }
             } else {
                 curPath.pop_back();
@@ -254,7 +259,7 @@ void PathsOutputWriter::addNode(common::nodeID_t nodeID, common::sel_t pos) cons
 DestinationsOutputWriter::DestinationsOutputWriter(main::ClientContext* context,
     RJOutputs* rjOutputs, processor::NodeOffsetMaskMap* outputNodeMask)
     : RJOutputWriter{context, rjOutputs, outputNodeMask} {
-    lengthVector = createVector(LogicalType::INT64(), context->getMemoryManager());
+    lengthVector = createVector(LogicalType::UINT16(), context->getMemoryManager());
 }
 
 void DestinationsOutputWriter::write(processor::FactorizedTable& fTable, nodeID_t dstNodeID) const {
@@ -262,7 +267,7 @@ void DestinationsOutputWriter::write(processor::FactorizedTable& fTable, nodeID_
         rjOutputs->ptrCast<SPOutputs>()->pathLengths->getMaskValueFromCurFrontierFixedMask(
             dstNodeID.offset);
     dstNodeIDVector->setValue<nodeID_t>(0, dstNodeID);
-    lengthVector->setValue<int64_t>(0, length);
+    setLength(lengthVector.get(), length);
     writeMoreAndAppend(fTable, dstNodeID, length);
 }
